Move 18FL to new structure
